+++ CHANGELOG.md
@@ -5,6 +5,11 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
 
 # Master (will become release 2.12)
 
+- Add `LagrangePrismLocalFiniteElement` with dynamic order, by generalizing the existing
+  static order implementation. The polynomial order of the shape functions is passed as constructor
+  parameter instead of a template parameter (which is implicitly set to `-1` as an indicator for the
+  dynamic polynomial order).
+
 - Add `LagrangePyramidLocalFiniteElement` with dynamic order, by generalizing the existing
   static order implementation. The polynomial order of the shape functions is passed as constructor
   parameter instead of a template parameter (which is implicitly set to `-1` as an indicator for the